Signup and Login: moved validation function into a created validation file
