+++ frontend/src/App.tsx
@@ -7,7 +7,6 @@ import Login from './pages/Login';
 function App() {
   return (
     <div className="App">
-    <Login/>
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
         <p>
